factored out CPP into its own file
