+++ parsley-core/parsley-core.cabal
@@ -78,7 +78,10 @@ library
                        Parsley.Internal.Backend.Machine.Types.State
 
   if impl(ghc >= 8.10)
-    exposed-modules:   Parsley.Internal.Backend.Machine.Types.Base,
+    exposed-modules:   Parsley.Internal.Backend.Machine.BindingOps,
+
+                       Parsley.Internal.Backend.Machine.Types.Base,
+                       Parsley.Internal.Backend.Machine.Types.Context,
                        Parsley.Internal.Backend.Machine.Types.Dynamics,
                        Parsley.Internal.Backend.Machine.Types.Offset,
                        Parsley.Internal.Backend.Machine.Types.Statics
