+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Ops.hs
@@ -2,56 +2,41 @@
 {-# LANGUAGE AllowAmbiguousTypes,
              ConstrainedClassMethods,
              ConstraintKinds,
-             CPP,
              ImplicitParams,
              MagicHash,
              NamedFieldPuns,
              PatternSynonyms,
              RecordWildCards,
              TypeApplications #-}
-module Parsley.Internal.Backend.Machine.Ops (module Parsley.Internal.Backend.Machine.Ops) where
+module Parsley.Internal.Backend.Machine.Ops (
+    module Parsley.Internal.Backend.Machine.Ops,
+    HandlerOps, JoinBuilder, RecBuilder, PositionOps, MarshalOps, LogOps
+  ) where
 
-import Control.Monad                                  (liftM2)
-import Control.Monad.Reader                           (ask, local)
-import Control.Monad.ST                               (ST)
-import Data.Array.Unboxed                             (UArray)
-import Data.ByteString.Internal                       (ByteString)
-import Data.STRef                                     (writeSTRef, readSTRef, newSTRef)
-import Data.Text                                      (Text)
-import Data.Void                                      (Void)
-import Debug.Trace                                    (trace)
-import GHC.Exts                                       (Int(..), (-#))
-import Language.Haskell.TH.Syntax                     (liftTyped)
-import Parsley.Internal.Backend.Machine.Defunc        (Defunc(OFFSET), genDefunc, _if, pattern FREEVAR)
-import Parsley.Internal.Backend.Machine.Identifiers   (MVar, ΦVar, ΣVar)
-import Parsley.Internal.Backend.Machine.InputOps      (PositionOps(..), LogOps(..), InputOps, next, more)
-import Parsley.Internal.Backend.Machine.InputRep      (Rep{-, representationTypes-})
-import Parsley.Internal.Backend.Machine.Instructions  (Access(..))
-import Parsley.Internal.Backend.Machine.LetBindings   (Regs(..))
-import Parsley.Internal.Backend.Machine.Types.Base
-import Parsley.Internal.Backend.Machine.Types.Dynamics
-import Parsley.Internal.Backend.Machine.Types.Offset  (Offset(..), moveOne, mkOffset)
-import Parsley.Internal.Backend.Machine.Types.State   {-(Γ(..), Ctx, Machine(..), MachineMonad, StaSubRoutine, OpStack(..), DynFunc, Cont#, Handler#,
-                                                      StaHandler(..), StaCont(..), DynHandler, DynCont, staHandler#, mkStaHandler, staCont#, mkStaCont, mkUnknown, staHandlerEval, unknown, mkFull,
-                                                      run, voidCoins, insertSub, insertΦ, insertNewΣ, cacheΣ, cachedΣ, concreteΣ, debugLevel,
-                                                      takeFreeRegisters,
-                                                      freshUnique, nextUnique)-}
+import Control.Monad                                   (liftM2)
+import Control.Monad.Reader                            (ask, local)
+import Control.Monad.ST                                (ST)
+import Data.STRef                                      (writeSTRef, readSTRef, newSTRef)
+import Data.Void                                       (Void)
+import Debug.Trace                                     (trace)
+import GHC.Exts                                        (Int(..), (-#))
+import Language.Haskell.TH.Syntax                      (liftTyped)
+import Parsley.Internal.Backend.Machine.BindingOps
+import Parsley.Internal.Backend.Machine.Defunc         (Defunc(OFFSET), genDefunc, _if, pattern FREEVAR)
+import Parsley.Internal.Backend.Machine.Identifiers    (MVar, ΦVar, ΣVar)
+import Parsley.Internal.Backend.Machine.InputOps       (PositionOps(..), LogOps(..), InputOps, next, more)
+import Parsley.Internal.Backend.Machine.InputRep       (Rep)
+import Parsley.Internal.Backend.Machine.Instructions   (Access(..))
+import Parsley.Internal.Backend.Machine.LetBindings    (Regs(..))
+import Parsley.Internal.Backend.Machine.Types          (MachineMonad, Machine(..), run)
+import Parsley.Internal.Backend.Machine.Types.Context
+import Parsley.Internal.Backend.Machine.Types.Dynamics (DynFunc, DynCont, DynHandler)
+import Parsley.Internal.Backend.Machine.Types.Offset   (Offset(..), moveOne, mkOffset)
+import Parsley.Internal.Backend.Machine.Types.State    (Γ(..), OpStack(..))
 import Parsley.Internal.Backend.Machine.Types.Statics
-import Parsley.Internal.Common                       (One, Code, Vec(..), Nat(..))
-import Parsley.Internal.Core.InputTypes              (Text16, CharList, Stream)
-import System.Console.Pretty                         (color, Color(Green, White, Red, Blue))
+import Parsley.Internal.Common                         (One, Code, Vec(..), Nat(..))
+import System.Console.Pretty                           (color, Color(Green, White, Red, Blue))
 
-import qualified Data.ByteString.Lazy.Internal as Lazy (ByteString)
-
-#define inputInstances(derivation) \
-derivation([Char])                 \
-derivation((UArray Int Char))      \
-derivation(Text16)                 \
-derivation(ByteString)             \
-derivation(CharList)               \
-derivation(Stream)                 \
-derivation(Lazy.ByteString)        \
-derivation(Text)
 
 type Ops o = (HandlerOps o, JoinBuilder o, RecBuilder o, PositionOps (Rep o), MarshalOps o, LogOps (Rep o))
 type StaHandlerBuilder s o a = Offset o -> StaHandler s o a
@@ -110,10 +95,10 @@ buildYesHandler :: Γ s o xs n r a
                 -> (Γ s o xs n r a -> Code (ST s (Maybe a)))
                 -> Word
                 -> StaHandler s o a
-buildYesHandler γ h u = StaHandler Nothing (mkUnknown $ \o# -> h (γ {input = mkOffset o# u})) Nothing
+buildYesHandler γ h u = staHandlerUnknown $ \o# -> h (γ {input = mkOffset o# u})
 
-fatal :: forall o s a. StaHandler s o a
-fatal = StaHandler Nothing (mkUnknown (const [|| returnST Nothing ||])) Nothing
+fatal :: StaHandler s o a
+fatal = staHandlerUnknown (const [|| returnST Nothing ||])
 
 bindAlwaysHandler :: HandlerOps o => Γ s o xs n r a
                   -> StaHandlerBuilder s o a
@@ -135,19 +120,6 @@ bindSameHandler γ yes no k = [||
                 k (γ {handlers = VCons (staHandlerFull (Just (input γ)) qhandler [||yesSame||] qno) (handlers γ)}))
   ||]
 
-class HandlerOps o where
-  bindHandler# :: StaHandler s o a -> (DynHandler s o a -> Code b) -> Code b
-
-#define deriveHandlerOps(_o)                                                                    \
-instance HandlerOps _o where                                                                    \
-{                                                                                               \
-  bindHandler# h k = [||                                                                        \
-    let handler (o# :: Rep _o) = $$(staHandler# h [||o#||])                                     \
-    in $$(k [||handler||])                                                                      \
-  ||];                                                                                          \
-};
-inputInstances(deriveHandlerOps)
-
 raise :: Γ s o xs (Succ n) r a -> Code (ST s (Maybe a))
 raise γ = let VCons h _ = handlers γ in staHandlerEval h (input γ)
 
@@ -155,14 +127,14 @@ raise γ = let VCons h _ = handlers γ in staHandlerEval h (input γ)
 suspend :: (Γ s o (x : xs) n r a -> Code (ST s (Maybe a))) -> Γ s o xs n r a -> Word -> StaCont s o a x
 suspend m γ u = mkStaCont $ \x o# -> m (γ {operands = Op (FREEVAR x) (operands γ), input = mkOffset o# u})
 
-halt :: forall o s a. StaCont s o a a
+halt :: StaCont s o a a
 halt = mkStaCont $ \x _ -> [||returnST (Just $$x)||]
 
-noreturn :: forall o s a. StaCont s o a Void
+noreturn :: StaCont s o a Void
 noreturn = mkStaCont $ \_ _ -> [||error "Return is not permitted here"||]
 
-callWithContinuation :: forall o s a x n. MarshalOps o => StaSubRoutine s o a x -> StaCont s o a x -> Code (Rep o) -> Vec (Succ n) (StaHandler s o a) -> Code (ST s (Maybe a))
-callWithContinuation sub ret input (VCons h _) = sub (dynCont @o ret) input (dynHandler @o h)
+callWithContinuation :: MarshalOps o => StaSubRoutine s o a x -> StaCont s o a x -> Code (Rep o) -> Vec (Succ n) (StaHandler s o a) -> Code (ST s (Maybe a))
+callWithContinuation sub ret input (VCons h _) = sub (dynCont ret) input (dynHandler h)
 
 resume :: StaCont s o a x -> Γ s o (x : xs) n r a -> Code (ST s (Maybe a))
 resume k γ = let Op x _ = operands γ in staCont# k (genDefunc x) (offset (input γ))
@@ -175,17 +147,6 @@ setupJoinPoint φ (Machine k) mx = freshUnique $ \u ->
         (\qx qo# -> mk (γ {operands = Op (FREEVAR qx) (operands γ), input = mkOffset qo# u}))
         (\qjoin -> run mx γ (insertΦ φ (staCont qjoin) ctx)))
       (local voidCoins k) ask
-
-class JoinBuilder o where
-  setupJoinPoint# :: StaCont# s o a x -> (DynCont s o a x -> Code b) -> Code b
-
-#define deriveJoinBuilder(_o)                                                             \
-instance JoinBuilder _o where                                                             \
-{                                                                                         \
-  setupJoinPoint# binding k =                                                             \
-    [|| let join x !(o# :: Rep _o) = $$(binding [||x||] [||o#||]) in $$(k [||join||]) ||] \
-};
-inputInstances(deriveJoinBuilder)
 
 buildIterAlways :: forall s o a. RecBuilder o => Ctx s o a -> MVar Void -> Machine s o '[] One Void a
                 -> StaHandlerBuilder s o a -> Offset o -> Word -> Code (ST s (Maybe a))
@@ -220,59 +181,14 @@ buildRec μ rs ctx k =
       run k (Γ Empty (staCont qret) (mkOffset qo# 0) (VCons (staHandler Nothing qh) VNil))
             (insertSub μ (\k o# h -> [|| $$qself $$k $$(o#) $$h ||]) (nextUnique ctx))
 
-class RecBuilder o where
-  bindIterHandler# :: (Code (Rep o) -> StaHandler# s o a)
-                   -> (Code (Rep o -> Handler# s o a) -> Code b)
-                   -> Code b
-  buildIter# :: Code (Rep o)
-             -> (Code (Rep o -> ST s (Maybe a)) -> Code (Rep o) -> Code (ST s (Maybe a)))
-             -> Code (ST s (Maybe a))
-  buildRec#  :: (DynSubRoutine s o a x -> DynCont s o a x -> Code (Rep o) -> DynHandler s o a -> Code (ST s (Maybe a)))
-             -> DynSubRoutine s o a x
-
-#define deriveRecBuilder(_o)                                                                           \
-instance RecBuilder _o where                                                                           \
-{                                                                                                      \
-  bindIterHandler# h k = [||                                                                           \
-      let handler (c# :: Rep _o) (o# :: Rep _o) = $$(h [||c#||] [||o#||]) in $$(k [||handler||])       \
-    ||];                                                                                               \
-  buildIter# o l = [||                                                                                 \
-      let loop !(o# :: Rep _o) = $$(l [||loop||] [||o#||])                                             \
-      in loop $$o                                                                                      \
-    ||];                                                                                               \
-  buildRec# binding =                                                                                  \
-    {- The idea here is to try and reduce the number of times registers have to be passed around -}    \
-    [|| let self ret !(o# :: Rep _o) h = $$(binding [||self||] [||ret||] [||o#||] [||h||]) in self ||] \
-};
-inputInstances(deriveRecBuilder)
-
 {- Marshalling Operations -}
-class MarshalOps o where
-  dynHandler :: StaHandler s o a -> DynHandler s o a
-  dynCont :: StaCont s o a x -> DynCont s o a x
+dynHandler :: forall s o a. MarshalOps o => StaHandler s o a -> DynHandler s o a
+dynHandler (StaHandler _ sh Nothing) = dynHandler# @o (unknown sh)
+dynHandler (StaHandler _ _ (Just dh)) = dh
 
-staHandler :: Maybe (Offset o) -> DynHandler s o a -> StaHandler s o a
-staHandler c dh = StaHandler c (mkUnknown (\o# -> [|| $$dh $$(o#) ||])) (Just dh)
-
-staHandlerFull :: Maybe (Offset o) -> DynHandler s o a -> Code (ST s (Maybe a)) -> DynHandler s o a -> StaHandler s o a
-staHandlerFull c handler yes no = StaHandler c
-  (mkFull (\o# -> [|| $$handler $$(o#) ||])
-          yes
-          (\o# -> [|| $$no $$(o#) ||]))
-  (Just handler)
-
-staCont :: DynCont s o a x -> StaCont s o a x
-staCont dk = StaCont (\x o# -> [|| $$dk $$x $$(o#) ||]) (Just dk)
-
-#define deriveMarshalOps(_o)                                                                  \
-instance MarshalOps _o where                                                                  \
-{                                                                                             \
-  dynHandler (StaHandler _ sh Nothing) = [||\ !(o# :: Rep _o) -> $$(unknown sh [||o#||]) ||]; \
-  dynHandler (StaHandler _ _ (Just dh)) = dh;                                                 \
-  dynCont (StaCont sk Nothing) = [||\ x (o# :: Rep _o) -> $$(sk [||x||] [||o#||]) ||];        \
-  dynCont (StaCont _ (Just dk)) = dk;                                                         \
-};
-inputInstances(deriveMarshalOps)
+dynCont :: forall s o a x. MarshalOps o => StaCont s o a x -> DynCont s o a x
+dynCont (StaCont sk Nothing) = dynCont# @o sk
+dynCont (StaCont _ (Just dk)) = dk
 
 {- Debugger Operations -}
 type LogHandler o = (PositionOps (Rep o), LogOps (Rep o))
@@ -299,9 +215,3 @@ preludeString name dir γ ctx ends = [|| concat [$$prelude, $$eof, ends, '\n' : 
     eof             = [|| if $$more $$end then $$inputTrace else $$inputTrace ++ color Red "•" ||]
     prelude         = [|| concat [indent, dir : name, dir : " (", show $$(offToInt offset), "): "] ||]
     caretSpace      = [|| replicate (length $$prelude + $$(offToInt offset) - $$(offToInt start)) ' ' ||]
-
--- RIP Dream :(
-{-$(let derive _o = [d|
-        instance HandlerOps _o where
-          fatal = [||\(!o#) -> return Nothing :: ST s (Maybe a)||]
-        |] in traverse derive representationTypes)-}
